test: exercise plugin close API
To help with API changes (and minimize the risk of accidental ones),
start building tests to explicitly exercise pieces of the plugin API,
and get more coverage in the module. Start with server->close().

Any plugin sources in test/plugins will be compiled and dropped into the
test server's modules directory by the Makefile's test recipes.